Add unsigned versions of ntohl and htonl
All systems i know of define ntohl and htonl to use unsigned 32bit
types. To keep backward compatibility for now I've added overloaded
versions so the impact on existing code is a small as possible. But I've
marked the signed versions as deprecated.

The code will now also use SwapEndian which is available as optimized
versions on some platforms.

git-svn-id: https://svn.freepascal.org/svn/fpc/trunk@22994 3ad0048d-3df7-0310-abae-a5850022a9f2